Add timeout for Consul requests
In OSXFUSE versions prior to 3.2.0, if the FUSE server takes too long to
respond to any request, the server will be considered dead: its mount
socket will be closed and all file system operations will start
returning errors until it is unmounted. To avoid this fatal error,
this commit adds a timeout to every request. The timeout is configurable
with a new "-timeout" option.

Even though this is a workaround for OS-X-specific behavior, the
timeouts apply to all OSs for consistency.

OSXFUSE versions from 3.2.0 onward don't need this behavior: a slow
request will instead be interrupted and does not immediately kill the
server. However, this version of OSXFUSE isn't considered "stable" yet,
so I'd rather work around the issue.