Convert dependencies to dep tool
Move from "godep" to "dep" for dependency management. It's going to be
the future.

Interestingly, "dep" insists on putting the entire repos of my
dependencies in the vendor directory (instead of just the packages
used). I'd rather not have all of Consul mirrored here, so "vendor/" is
removed. I suppose this also makes it easier to import the filesystem as
a dependency.