Update dependencies
In particular, updates to the latest iteration of the Consul API
package, which is now its own module.

This update to bazil.org/fuse is likely the last. Past this version, the
"allow_root" option has been removed, and as of April 2020, macOS
support has been removed completely. This is a no-go for me personally,
though with FUSE for macOS going closed-source and Apple locking down
kernel extensions, I can understand.